requirements.txt
beautifulsoup4
langchain
langgraph>=0.0.25
langchainhub
langchain-community
langchain-huggingface
tavily-python
python-dotenv
black
isort
pytest
llama-parse @ git+https://github.com/run-llama/llama_parse.git
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
streamlit>=1.24.0
sentence-transformers
torch>=2.0.0

